Renamed and moved some tests after last commit
Renamed `test_course.py` to `test_course_forms.py`.

Renamed `test_reservation_rules.py` to `test_reservation_rule.py`.

Moved from `test_reservation_reservation.py` to `test_api_reservation_views.py`:
* TestAPIReservationMarkFinishedView
Moved from `test_delete_reservation_view.py` to `test_api_reservation_views.py`:
* TestAPIReservationDeleteView

Renamed `test_reservation_reservation.py` to `test_reservation_views.py`.
