+++ src/make_queue/tests/views/test_api_reservation_views.py
@@ -1,9 +1,12 @@
 from datetime import datetime, timedelta
 from http import HTTPStatus
 from unittest import mock
+from unittest.mock import patch
 
 from django.contrib.auth.models import AnonymousUser
 from django.test import Client, TestCase
+from django.utils import timezone
+from django.utils.dateparse import parse_time
 from django.utils.http import urlencode
 from django_hosts import reverse
 
@@ -13,7 +16,11 @@ from util.locale_utils import parse_datetime_localized
 from ...api.views import APIReservationListView
 from ...models.course import Printer3DCourse
 from ...models.machine import Machine, MachineType
-from ...models.reservation import Quota, Reservation
+from ...models.reservation import Quota, Reservation, ReservationRule
+from ...templatetags.reservation_extra import can_change_reservation
+
+
+Day = ReservationRule.Day
 
 
 class APIReservationListViewTests(TestCase):
@@ -156,3 +163,190 @@ class APIReservationListViewTests(TestCase):
             'field_errors': {'end_date': [field_required]}, **undefined_asdf_field})
         assert_error_response_contains("end_date=2020-01&asdf=asdf", expected_json_dict={
             'field_errors': {'start_date': [field_required], 'end_date': [value_not_datetime]}, **undefined_asdf_field})
+
+
+class TestAPIReservationMarkFinishedView(TestCase):
+
+    def setUp(self):
+        self.user = User.objects.create_user("test")
+        self.client.force_login(self.user)
+        # See the `0015_machinetype.py` migration for which MachineTypes are created by default
+        self.machine_type = MachineType.objects.get(pk=2)
+        self.machine = Machine.objects.create(machine_type=self.machine_type, status=Machine.Status.AVAILABLE, name="Test")
+        Quota.objects.create(machine_type=self.machine_type, number_of_reservations=2, ignore_rules=False, all=True)
+        ReservationRule.objects.create(
+            machine_type=self.machine_type, start_time=parse_time("00:00"), days_changed=6, end_time=parse_time("23:59"),
+            start_days=[Day.MONDAY], max_hours=6, max_inside_border_crossed=6,
+        )
+        self.now = timezone.localtime()
+        self.reservation1 = Reservation.objects.create(
+            machine=self.machine, user=self.user,
+            start_time=self.now + timedelta(hours=1),
+            end_time=self.now + timedelta(hours=2),
+        )
+
+    def post_to(self, reservation: Reservation):
+        return self.client.post(reverse('api_reservation_mark_finished', args=[reservation.pk]))
+
+    def test_get_request_fails(self):
+        response = self.client.get(reverse('api_reservation_mark_finished', args=[self.reservation1.pk]))
+        self.assertEqual(response.status_code, HTTPStatus.NOT_FOUND)
+
+    @patch('django.utils.timezone.now')
+    def test_valid_post_request_succeeds(self, now_mock):
+        # Freeze the return value of `timezone.now()` and set it to 1 minute after `self.reservation1` has started
+        now_mock.return_value = self.now + timedelta(hours=1, minutes=1)
+        self.assertTrue(can_change_reservation(self.reservation1, self.user))
+
+        response = self.post_to(self.reservation1)
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        self.reservation1.refresh_from_db()
+        self.assertEqual(self.reservation1.end_time, timezone.now())
+
+    def test_finishing_before_start_fails(self):
+        original_end_time = self.reservation1.end_time
+        response = self.post_to(self.reservation1)
+        self.assertEqual(response.status_code, HTTPStatus.BAD_REQUEST)
+        self.reservation1.refresh_from_db()
+        self.assertEqual(self.reservation1.end_time, original_end_time,
+                         "Marking a reservation in the future as finished should not be possible")
+
+    @patch('django.utils.timezone.now')
+    def test_finishing_after_end_fails(self, now_mock):
+        # Set "now" to 1 hour after `self.reservation1` has ended
+        now_mock.return_value = self.now + timedelta(hours=3)
+
+        original_end_time = self.reservation1.end_time
+        response = self.post_to(self.reservation1)
+        self.assertEqual(response.status_code, HTTPStatus.BAD_REQUEST)
+        self.reservation1.refresh_from_db()
+        self.assertEqual(self.reservation1.end_time, original_end_time,
+                         "Marking a reservation in the past as finished should not do anything")
+
+    @patch('django.utils.timezone.now')
+    def test_finishing_just_before_other_reservation_starts_succeeds(self, now_mock):
+        # Freeze the return value of `timezone.now()`
+        now_mock.return_value = self.now
+
+        self.reservation1.delete()
+        reservation2 = Reservation.objects.create(
+            machine=self.machine, user=self.user,
+            start_time=self.now + timedelta(minutes=1),
+            end_time=self.now + timedelta(hours=6),
+        )
+        reservation3 = Reservation.objects.create(
+            machine=self.machine, user=User.objects.create_user("test2"),
+            start_time=self.now + timedelta(hours=6),
+            end_time=self.now + timedelta(hours=6, minutes=26),
+        )
+
+        # Set "now" to 3 minutes before `reservation2` ends and `reservation3` starts
+        now_mock.return_value = self.now + timedelta(hours=5, minutes=57)
+        response = self.post_to(reservation2)
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        reservation2.refresh_from_db()
+        self.assertEqual(reservation2.end_time, timezone.now())
+
+    @patch('django.utils.timezone.now')
+    def test_can_finish_existing_reservations_for_machine_out_of_order_or_on_maintenance(self, now_mock):
+        # Freeze the return value of `timezone.now()`
+        now_mock.return_value = self.now
+
+        for machine_status in (Machine.Status.OUT_OF_ORDER, Machine.Status.MAINTENANCE):
+            with self.subTest(machine_status=machine_status):
+                machine = self.reservation1.machine
+                machine.status = machine_status
+                machine.save()
+
+                # Set "now" to 5 minutes before `self.reservation1` ends
+                now_mock.return_value = self.reservation1.end_time - timedelta(minutes=5)
+
+                response = self.post_to(self.reservation1)
+                self.assertEqual(response.status_code, HTTPStatus.OK)
+                self.reservation1.refresh_from_db()
+                self.assertEqual(self.reservation1.end_time, timezone.now())
+
+
+class TestAPIReservationDeleteView(TestCase):
+
+    def setUp(self):
+        password = "weak_pass"
+        self.user1 = User.objects.create_user("user1", "user1@makentnu.no", password)
+        self.user2 = User.objects.create_user("user2", "user2@makentnu.no", password)
+        self.client1 = Client()
+        self.client2 = Client()
+        self.client1.login(username=self.user1.username, password=password)
+        self.client2.login(username=self.user2.username, password=password)
+
+        # See the `0015_machinetype.py` migration for which MachineTypes are created by default
+        printer_machine_type = MachineType.objects.get(pk=1)
+        self.machine1 = Machine.objects.create(
+            name="U1", machine_model="Ultimaker 2", machine_type=printer_machine_type,
+            location="MAKE", status=Machine.Status.AVAILABLE,
+        )
+
+        Quota.objects.create(user=self.user1, number_of_reservations=2, ignore_rules=True, machine_type=printer_machine_type)
+        Quota.objects.create(user=self.user2, number_of_reservations=2, ignore_rules=True, machine_type=printer_machine_type)
+
+        now = timezone.localtime()
+        Printer3DCourse.objects.create(user=self.user1, username=self.user1.username, name=self.user1.get_full_name(), date=now)
+        Printer3DCourse.objects.create(user=self.user2, username=self.user2.username, name=self.user2.get_full_name(), date=now)
+
+        self.reservation = Reservation.objects.create(
+            user=self.user1,
+            machine=self.machine1,
+            start_time=now + timedelta(hours=2),
+            end_time=now + timedelta(hours=4),
+        )
+
+    def test_delete_own_reservation_succeeds(self):
+        response = self.client1.delete(reverse('api_reservation_delete', args=[self.reservation.pk]))
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        self.assertFalse(Reservation.objects.filter(pk=self.reservation.pk).exists())
+
+    def test_delete_other_users_reservation_fails(self):
+        response = self.client2.delete(reverse('api_reservation_delete', args=[self.reservation.pk]))
+        self.assertEqual(response.status_code, HTTPStatus.FORBIDDEN)
+        self.assertTrue(Reservation.objects.filter(pk=self.reservation.pk).exists())
+
+    def test_delete_only_one_among_multiple_reservations_succeeds(self):
+        now = timezone.localtime()
+        reservation2 = Reservation.objects.create(
+            user=self.user1,
+            machine=self.machine1,
+            start_time=now + timedelta(hours=6),
+            end_time=now + timedelta(hours=8),
+        )
+
+        response = self.client1.delete(reverse('api_reservation_delete', args=[self.reservation.pk]))
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        self.assertFalse(Reservation.objects.filter(pk=self.reservation.pk).exists())
+        self.assertTrue(Reservation.objects.filter(pk=reservation2.pk).exists())
+
+    def test_delete_reservation_after_start_fails(self):
+        now = timezone.localtime()
+        Reservation.objects.filter(pk=self.reservation.pk).update(start_time=now)
+        response = self.client1.delete(reverse('api_reservation_delete', args=[self.reservation.pk]))
+        self.assertEqual(response.status_code, HTTPStatus.BAD_REQUEST)
+        self.assertTrue(Reservation.objects.filter(pk=self.reservation.pk).exists())
+
+        # Setting the start time to the future should allow deletion
+        Reservation.objects.filter(pk=self.reservation.pk).update(start_time=now + timedelta(hours=1))
+        response = self.client1.delete(reverse('api_reservation_delete', args=[self.reservation.pk]))
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        self.assertFalse(Reservation.objects.filter(pk=self.reservation.pk).exists())
+
+    def test_can_cancel_existing_reservation_for_machine_out_of_order(self):
+        self._test_can_cancel_existing_reservation_for_machine_with_status(Machine.Status.OUT_OF_ORDER)
+
+    def test_can_cancel_existing_reservation_for_machine_on_maintenance(self):
+        self._test_can_cancel_existing_reservation_for_machine_with_status(Machine.Status.MAINTENANCE)
+
+    def _test_can_cancel_existing_reservation_for_machine_with_status(self, machine_status: Machine.Status):
+        machine = self.reservation.machine
+        machine.status = machine_status
+        machine.save()
+
+        response = self.client1.delete(reverse('api_reservation_delete', args=[self.reservation.pk]))
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        self.assertFalse(Reservation.objects.filter(pk=self.reservation.pk).exists())
